src/main/java/me/o_nix/assignments/currency_pairs/CurrencyPairsApplication.java
package me.o_nix.assignments.currency_pairs;

import com.github.marschall.threeten.jpa.LocalDateTimeConverter;
import org.apache.http.client.HttpClient;
import org.apache.http.impl.client.HttpClientBuilder;
import org.springframework.boot.SpringApplication;
import org.springframework.boot.autoconfigure.SpringBootApplication;
import org.springframework.boot.orm.jpa.EntityScan;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.EnableAspectJAutoProxy;
import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
import org.springframework.http.client.ClientHttpRequestFactory;
import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
import org.springframework.transaction.annotation.EnableTransactionManagement;
import org.springframework.web.client.RestTemplate;

@SpringBootApplication
@EnableJpaRepositories
@EnableTransactionManagement
@EnableAspectJAutoProxy
@EntityScan(basePackageClasses = {
		CurrencyPairsApplication.class,
		LocalDateTimeConverter.class})
public class CurrencyPairsApplication {
	public static void main(String[] args) {
		SpringApplication.run(CurrencyPairsApplication.class, args);
	}

	@Bean
	public RestTemplate restTemplate() {
		HttpClient httpClient = HttpClientBuilder.create().build();
		ClientHttpRequestFactory requestFactory = new HttpComponentsClientHttpRequestFactory(httpClient);

		return new RestTemplate(requestFactory);
	}
}